requirements.txt
aio-mc-rcon==3.0.2
aiofiles==0.7.0
aiogTTS==1.1.0
aiohttp==3.7.4.post0
aiomysql==0.0.21
astunparse==1.6.3
async-timeout==3.0.1
asyncio-dgram==2.0.0
asyncpg==0.23.0
asyncrcon==1.1.4
attrs==21.2.0
Automat==0.8.0
beautifulsoup4==4.9.3
blinker==1.4
botocore==1.20.66
braceexpand==0.1.7
certifi==2019.11.28
cffi==1.14.6
chardet==4.0.0
cleverbotfreeapi==1.1.0
click==8.0.1
cloud-init==21.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
DateTime==4.3
dbus-python==1.2.16
discord==1.7.3
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@309c70222eec33acacb8c21bbb0fbe9bc57de661
discord-ext-menus-views @ git+https://github.com/oliver-ni/discord-ext-menus-views@cd1d6f89e69299d49e886e1e148256640bd6bfcf
discord-timers==0.1.0
discord.py @ git+https://github.com/Rapptz/discord.py@fbc4a51c355c029af61363038d0a8a9ac0bd8ace
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.1.0
ec2-hibinit-agent==1.0.0
ec2instanceconnectcli==1.0.2
entrypoints==0.3
ffmpeg-python==0.2.0
future==0.18.2
gTTS==2.2.3
hibagent==1.0.1
httplib2==0.14.0
humanize==3.7.1
hyperlink==19.0.0
idna==3.2
import-expression==1.1.4
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jishaku==2.2.0
jmespath==0.10.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
mcstatus==6.0.0
more-itertools==4.2.0
motor==2.4.0
multidict==5.1.0
netifaces==0.10.4
oauthlib==3.1.0
parsedatetime==2.6
pexpect==4.6.0
Pillow==8.2.0
psutil==5.8.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.12.0
PyMySQL==0.9.3
PyNaCl==1.4.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-unixsocket==0.2.0
roblox.py==0.2.4
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sos==4.1
soupsieve==2.2.1
ssh-import-id==5.10
supervisor==4.2.2
systemd-python==234
Twisted==18.9.0
typing==3.7.4.3
typing-extensions==3.10.0.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.2.0
urllib3==1.26.5
wadllib==1.3.3
wavelink==1.0.0b30
yarl==1.6.3
youtube-dl==2021.6.6
zipp==3.4.1
zope.interface==5.4.0
